refaktoroitu fronend
